requirements.txt
PyQt6>=6.10
Pillow>=10.0
